src/main/java/hello/jdbc/connection/DBConnectionUtil.java
package hello.jdbc.connection;

import lombok.extern.slf4j.Slf4j;

import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.SQLException;

//static import로 상수 사용
import static hello.jdbc.connection.ConnectionConst.*;

@Slf4j
public class DBConnectionUtil {

    public static Connection getConnection(){
        try {
            Connection connection = DriverManager.getConnection(URL, USERNAME, PASSWORD);
            log.info("get connection={}, class={}", connection, connection.getClass());
            return connection;
        } catch (SQLException e) {
            //stacktrace 하지 말고 RuntimeException으로 변경해서 던짐.
            throw new IllegalStateException(e);
        }
    }
}
